+++ exporters/otlp/otlptrace/otlptracegrpc/client_test.go
@@ -19,7 +19,11 @@ import (
 	"google.golang.org/grpc/backoff"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/encoding/gzip"
+	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
+
+	coltracepb "go.opentelemetry.io/proto/otlp/collector/trace/v1"
+	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
@@ -28,8 +32,6 @@ import (
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc/internal/otlptracetest"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
-	coltracepb "go.opentelemetry.io/proto/otlp/collector/trace/v1"
-	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 )
 
 func TestMain(m *testing.M) {
@@ -216,6 +218,8 @@ func TestNewWithHeaders(t *testing.T) {
 	t.Cleanup(func() { require.NoError(t, mc.stop()) })
 
 	ctx := context.Background()
+	additionalKey := "additional-custom-header"
+	ctx = metadata.AppendToOutgoingContext(ctx, additionalKey, "additional-value")
 	exp := newGRPCExporter(t, ctx, mc.endpoint,
 		otlptracegrpc.WithHeaders(map[string]string{"header1": "value1"}))
 	t.Cleanup(func() { require.NoError(t, exp.Shutdown(ctx)) })
@@ -224,6 +228,7 @@ func TestNewWithHeaders(t *testing.T) {
 	headers := mc.getHeaders()
 	require.Regexp(t, "OTel OTLP Exporter Go/1\\..*", headers.Get("user-agent"))
 	require.Len(t, headers.Get("header1"), 1)
+	require.Len(t, headers.Get(additionalKey), 1)
 	assert.Equal(t, "value1", headers.Get("header1")[0])
 }